backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
openai
huggingface_hub
